requirement.txt
# requirements.txt

tensorflow
scikit-learn
pandas
numpy
Flask
python-dotenv
google-generativeai
gunicorn